src/utils_v3/docker/requirements.txt
pyyaml
gitignore-parser
dockerfile-parse
